requirements.txt
asgiref==3.8.1
boto3==1.34.131
botocore==1.34.131
Django==5.0.6
django-cors-headers==4.3.1
django-filter==24.2
django-phonenumber-field==7.3.0
django-rest-passwordreset==1.4.1
django-storages==1.14.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
environs==11.0.0
jmespath==1.0.1
marshmallow==3.21.2
packaging==24.0
phonenumberslite==8.13.37
pillow==10.3.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
s3transfer==0.10.1
six==1.16.0
sqlparse==0.5.0
text-unidecode==1.3
urllib3==2.2.2
